Updated with Mark Kilgard's changes

+++ lib/mui/browseparse.c
@@ -80,7 +80,7 @@ void parsebrowsefile(FILE *f)
 
 void setcurrentfilename(char *s)
 {
-    int len = (int)(strlen(s));
+    int len = strlen(s);
     char *sptr;
     
     sptr = &s[len-1];
